[Refactor] : MovieRepository 리펙토링 #10
